requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6  # For handling file uploads

# AI and NLP
openai==1.3.5
langchain==0.0.350
faiss-cpu==1.7.4
openai-whisper==20231117
sentence-transformers==2.2.2

# Audio Processing
pyttsx3==2.90
pydub==0.25.1
sounddevice==0.4.6
scipy==1.11.4

# Environment and Utilities
python-dotenv==1.0.0
pydantic==2.4.2
numpy==1.26.2
jinja2==3.1.2
aiofiles==23.2.1
torch==2.1.1 